src/tensors/accelerator_backends/opengl_pipeline.py
#!/usr/bin/env python3
"""Prototype classes for OpenGL shader orchestration."""
from __future__ import annotations

try:
    from typing import Any, Dict, Tuple
    from pathlib import Path
    from OpenGL import GL  # type: ignore
except Exception:
    import sys
    print("OpenGL pipeline failed to import")
    sys.exit(1)
# --- END HEADER ---


# ########## STUB: OpenGL Shader Pipeline ##########
# PURPOSE: Provide a composable system that manages compute shader source
#          fragments, compiles them, and caches the resulting programs for
#          efficient reuse. The pipeline will also manage OpenGL buffer objects
#          used to store tensor data during dynamic operations.
# EXPECTED BEHAVIOR: When implemented, this module will allow creation of
#          operation chains by stitching together shader snippet middles. The
#          `BatchDelegator` will decide when to flush a chain based on a batch
#          size cap and reconfigure the pipeline for continued execution.
# INPUTS: GLSL snippet paths or strings, numpy staging buffers, and runtime
#         parameters describing tensor shapes and workgroup sizes.
# OUTPUTS: Compiled ``GLuint`` program handles and bound buffer objects ready
#          for dispatch.
# KEY ASSUMPTIONS/DEPENDENCIES: Requires a valid OpenGL context and the
#         ``PyOpenGL`` package. Assumes shader snippets use the INIT/OPERATIONS/
#         OUTPUT sentinels documented in ``glsl_kernels/AGENTS.md``.
# TODO:
#   - Implement shader composition and compilation.
#   - Track and reuse compiled programs via ``ShaderCache``.
#   - Manage buffer pooling and resizing in ``GLBufferManager``.
#   - Add synchronization utilities for reading results back to CPU memory.
# NOTES: This stub defines class interfaces only. All methods raise
#         ``NotImplementedError`` to highlight missing functionality.
# ###########################################################################


class GLBufferManager:
    """Allocate and recycle OpenGL buffer objects."""

    def __init__(self) -> None:
        self._buffers: Dict[int, int] = {}

    def allocate(self, size: int) -> int:
        """Return a buffer id with at least ``size`` bytes."""
        raise NotImplementedError

    def release(self, buf_id: int) -> None:
        """Free a buffer from internal tracking and delete it."""
        raise NotImplementedError

    def resize(self, buf_id: int, new_size: int) -> None:
        """Resize an existing buffer object."""
        raise NotImplementedError


class ShaderCompiler:
    """Compile GLSL compute shader sources."""

    @staticmethod
    def compile_source(src: str) -> int:
        """Compile ``src`` into a program and return its id."""
        raise NotImplementedError

    @staticmethod
    def load_and_compile(path: Path) -> int:
        return ShaderCompiler.compile_source(path.read_text())


class ShaderCache:
    """Cache compiled shader programs keyed by source hash."""

    def __init__(self) -> None:
        self._programs: Dict[str, int] = {}

    def get(self, key: str) -> int | None:
        return self._programs.get(key)

    def add(self, key: str, program: int) -> None:
        self._programs[key] = program


class ShaderComposer:
    """Combine shader snippets using sentinel markers."""

    def __init__(self, kernel_dir: Path) -> None:
        self.kernel_dir = kernel_dir

    def compose(self, names: list[str]) -> str:
        """Return a full shader source composed from snippets ``names``."""
        raise NotImplementedError


class BatchDelegator:
    """Manage operation chains and batch limits."""

    def __init__(self, max_ops: int = 4) -> None:
        self.max_ops = max_ops
        self.current_ops: list[str] = []
        self.cache = ShaderCache()
        self.compiler = ShaderCompiler()
        self.composer = ShaderComposer(Path(__file__).with_name("glsl_kernels"))
        self.buffers = GLBufferManager()

    def append_operation(self, name: str) -> None:
        """Add a shader operation and compile batch if limit reached."""
        raise NotImplementedError

    def dispatch(self, buffer_id: int, size: Tuple[int, ...]) -> None:
        """Execute the currently composed program on ``buffer_id``."""
        raise NotImplementedError

